+++ Src/DelphiCompileGate.BuildInfo.pas
@@ -8,7 +8,7 @@ const
   DCG_PROTOCOL_VERSION = 2;
   DCG_PLUGIN_VERSION = '2.0.0';
   DCG_IDE_VERSION = 'Delphi 13 Community Edition';
-  DCG_PACKAGE_BUILD_ID = 'dcg-v2-20260904-releaseprep-02';
+  DCG_PACKAGE_BUILD_ID = 'dcg-v2-20260905-searchpathfix-03';
 
 implementation